:truck notification으로 이름 변경
